requirements.txt
Flask==2.3.3
pytube2==15.0.16
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
Werkzeug==2.3.7